LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpTransceiver-stopping.https_rest-expected.txt


PASS [audio] Locally stopping a transceiver ends the track
FAIL [audio] Remotely stopping a transceiver ends the track assert_equals: expected "live" but got "ended"
PASS [audio] Rollback when transceiver is not removed does not end track
FAIL [audio] Rollback when removing transceiver does end the track assert_equals: expected "live" but got "ended"
PASS [audio] Glare when transceiver is not removed does not end track
PASS [video] Locally stopping a transceiver ends the track
FAIL [video] Remotely stopping a transceiver ends the track assert_equals: expected "live" but got "ended"
PASS [video] Rollback when transceiver is not removed does not end track
FAIL [video] Rollback when removing transceiver does end the track assert_equals: expected "live" but got "ended"
PASS [video] Glare when transceiver is not removed does not end track
